refactor: 가격 타입을 Long으로 변경 및 장바구니 상품 수량 제한 추가

+++ app/src/main/java/woowacourse/shopping/core/uimodel/CartItemUiModel.kt
@@ -3,5 +3,5 @@ package woowacourse.shopping.core.uimodel
 data class CartItemUiModel(
     val product: ProductUiModel,
     val quantity: Int,
-    val totalPrice: Int,
+    val totalPrice: Long,
 )
